refactor(api-headless-cms): inline entry data factory logic, wire use cases to inject factories
