--- packages/api-headless-cms/src/crud/contentEntry/entryDataFactories/createEntryRevisionFromData.ts
@@ -1,223 +1,0 @@
-import type {
-    CmsContext,
-    CmsEntry,
-    CmsEntryValues,
-    CmsModel,
-    CreateCmsEntryInput,
-    CreateCmsEntryOptionsInput
-} from "~/types/index.js";
-import { getDate } from "~/utils/date.js";
-import { getIdentity } from "~/utils/identity.js";
-import { mapAndCleanUpdatedInputData } from ".//mapAndCleanUpdatedInputData.js";
-import { validateModelEntryDataOrThrow } from "../entryDataValidation.js";
-import { referenceFieldsMapping } from "../referenceFieldsMapping.js";
-import { createIdentifier, parseIdentifier } from "@webiny/utils";
-import WebinyError from "@webiny/error";
-import type { SecurityIdentity } from "@webiny/api-core/types/security.js";
-import type { Tenant } from "@webiny/api-core/types/tenancy.js";
-import { STATUS_DRAFT, STATUS_PUBLISHED, STATUS_UNPUBLISHED } from "./statuses.js";
-import type { AccessControl } from "~/crud/AccessControl/AccessControl.js";
-import { NotAuthorizedError } from "~/utils/errors.js";
-import { getSystem } from "./system.js";
-
-interface CreateEntryRevisionFromDataParams<TValues extends CmsEntryValues = CmsEntryValues> {
-    sourceId: string;
-    model: CmsModel;
-    rawInput: CreateCmsEntryInput<TValues>;
-    options?: CreateCmsEntryOptionsInput;
-    context: CmsContext;
-    getIdentity: () => SecurityIdentity;
-    getTenant: () => Tenant;
-    originalEntry: CmsEntry<TValues>;
-    latestStorageEntry: CmsEntry<TValues>;
-    accessControl: AccessControl;
-}
-
-interface ICreateEntryRevisionFromDataResponse<TValues extends CmsEntryValues = CmsEntryValues> {
-    entry: CmsEntry<TValues>;
-    input: CreateCmsEntryInput<TValues>;
-}
-
-export const createEntryRevisionFromData = async <TValues extends CmsEntryValues = CmsEntryValues>({
-    sourceId,
-    model,
-    rawInput,
-    options,
-    context,
-    getIdentity: getSecurityIdentity,
-    originalEntry,
-    latestStorageEntry,
-    accessControl
-}: CreateEntryRevisionFromDataParams<TValues>): Promise<
-    ICreateEntryRevisionFromDataResponse<TValues>
-> => {
-    /**
-     * Make sure we only work with fields that are defined in the model.
-     *
-     * Also, we can be certain to cast as TValues because originalValues and new ones are being merged.
-     */
-    const initialValues = {
-        ...originalEntry.values,
-        ...mapAndCleanUpdatedInputData<TValues>(model, rawInput.values)
-    };
-
-    await validateModelEntryDataOrThrow({
-        context,
-        model,
-        values: initialValues,
-        entry: originalEntry,
-        skipValidators: options?.skipValidators
-    });
-
-    const values = await referenceFieldsMapping<TValues>({
-        context,
-        model,
-        values: initialValues,
-        validateEntries: false
-    });
-
-    const latestId = latestStorageEntry ? latestStorageEntry.id : sourceId;
-    const { id, version: nextVersion } = increaseEntryIdVersion(latestId);
-
-    const currentIdentity = getSecurityIdentity();
-    const currentDateTime = new Date();
-
-    /**
-     * Users can set the initial status of the entry. If so, we need to make
-     * sure they have the required permissions and also that all the fields
-     * are filled in correctly.
-     */
-    const status = rawInput.status || STATUS_DRAFT;
-    if (status !== STATUS_DRAFT) {
-        if (status === STATUS_PUBLISHED) {
-            const canPublish = await accessControl.canAccessEntry({ model, pw: "p" });
-            if (!canPublish) {
-                throw new NotAuthorizedError(`Not allowed to access "${model.modelId}" entries.`);
-            }
-        } else if (status === STATUS_UNPUBLISHED) {
-            const canUnpublish = await accessControl.canAccessEntry({ model, pw: "u" });
-            if (!canUnpublish) {
-                throw new NotAuthorizedError(`Not allowed to access "${model.modelId}" entries.`);
-            }
-        }
-    }
-
-    const locked = status !== STATUS_DRAFT;
-
-    let revisionLevelPublishingMetaFields: Pick<
-        CmsEntry,
-        | "revisionFirstPublishedOn"
-        | "revisionLastPublishedOn"
-        | "revisionFirstPublishedBy"
-        | "revisionLastPublishedBy"
-    > = {
-        revisionFirstPublishedOn: getDate(rawInput.revisionFirstPublishedOn, null),
-        revisionLastPublishedOn: getDate(rawInput.revisionLastPublishedOn, null),
-        revisionFirstPublishedBy: getIdentity(rawInput.revisionFirstPublishedBy, null),
-        revisionLastPublishedBy: getIdentity(rawInput.revisionLastPublishedBy, null)
-    };
-
-    let entryLevelPublishingMetaFields: Pick<
-        CmsEntry,
-        "firstPublishedOn" | "lastPublishedOn" | "firstPublishedBy" | "lastPublishedBy"
-    > = {
-        firstPublishedOn: getDate(rawInput.firstPublishedOn, latestStorageEntry.firstPublishedOn),
-        lastPublishedOn: getDate(rawInput.lastPublishedOn, latestStorageEntry.lastPublishedOn),
-        firstPublishedBy: getIdentity(
-            rawInput.firstPublishedBy,
-            latestStorageEntry.firstPublishedBy
-        ),
-        lastPublishedBy: getIdentity(rawInput.lastPublishedBy, latestStorageEntry.lastPublishedBy)
-    };
-
-    if (status === STATUS_PUBLISHED) {
-        revisionLevelPublishingMetaFields = {
-            revisionFirstPublishedOn: getDate(rawInput.revisionFirstPublishedOn, currentDateTime),
-            revisionLastPublishedOn: getDate(rawInput.revisionLastPublishedOn, currentDateTime),
-            revisionFirstPublishedBy: getIdentity(
-                rawInput.revisionFirstPublishedBy,
-                currentIdentity
-            ),
-            revisionLastPublishedBy: getIdentity(rawInput.revisionLastPublishedBy, currentIdentity)
-        };
-
-        entryLevelPublishingMetaFields = {
-            firstPublishedOn: getDate(
-                rawInput.firstPublishedOn,
-                latestStorageEntry.firstPublishedOn
-            ),
-            lastPublishedOn: getDate(rawInput.lastPublishedOn, currentDateTime),
-            firstPublishedBy: getIdentity(
-                rawInput.firstPublishedBy,
-                latestStorageEntry.firstPublishedBy
-            ),
-            lastPublishedBy: getIdentity(rawInput.lastPublishedBy, currentIdentity)
-        };
-    }
-
-    const entry: CmsEntry<TValues> = {
-        ...originalEntry,
-        id,
-        version: nextVersion,
-
-        /**
-         * Entry-level meta fields. 👇
-         */
-        createdOn: getDate(rawInput.createdOn, latestStorageEntry.createdOn),
-        savedOn: getDate(rawInput.savedOn, currentDateTime),
-        modifiedOn: getDate(rawInput.modifiedOn, currentDateTime),
-        createdBy: getIdentity(rawInput.createdBy, latestStorageEntry.createdBy)!,
-        savedBy: getIdentity(rawInput.savedBy, currentIdentity)!,
-        modifiedBy: getIdentity(rawInput.modifiedBy, currentIdentity),
-        ...entryLevelPublishingMetaFields,
-
-        /**
-         * Revision-level meta fields. 👇
-         */
-        revisionCreatedOn: getDate(rawInput.revisionCreatedOn, currentDateTime),
-        revisionSavedOn: getDate(rawInput.revisionSavedOn, currentDateTime),
-        revisionModifiedOn: getDate(rawInput.revisionModifiedOn, null),
-        revisionCreatedBy: getIdentity(rawInput.revisionCreatedBy, currentIdentity)!,
-        revisionSavedBy: getIdentity(rawInput.revisionSavedBy, currentIdentity)!,
-        revisionModifiedBy: getIdentity(rawInput.revisionModifiedBy, null),
-        ...revisionLevelPublishingMetaFields,
-
-        locked,
-        status,
-        values,
-        system: getSystem({
-            input: rawInput,
-            original: originalEntry
-        }),
-        live: originalEntry.live
-    };
-
-    return {
-        entry,
-        input: {
-            ...rawInput,
-            values: structuredClone(values)
-        }
-    };
-};
-
-const increaseEntryIdVersion = (id: string) => {
-    const { id: entryId, version } = parseIdentifier(id);
-    if (!version) {
-        throw new WebinyError(
-            "Cannot increase version on the ID without the version part.",
-            "WRONG_ID",
-            {
-                id
-            }
-        );
-    }
-    return {
-        entryId,
-        version: version + 1,
-        id: createIdentifier({
-            id: entryId,
-            version: version + 1
-        })
-    };
-};
