--- packages/api-headless-cms/src/crud/contentEntry/entryDataFactories/createUpdateEntryData.ts
@@ -1,186 +1,0 @@
-import type {
-    CmsContext,
-    CmsEntry,
-    CmsEntryStatus,
-    CmsEntryValues,
-    CmsModel,
-    UpdateCmsEntryInput,
-    UpdateCmsEntryOptionsInput
-} from "~/types/index.js";
-import { getDate } from "~/utils/date.js";
-import { getIdentity } from "~/utils/identity.js";
-import { validateModelEntryDataOrThrow } from "../entryDataValidation.js";
-import { referenceFieldsMapping } from "../referenceFieldsMapping.js";
-import { mapAndCleanUpdatedInputData } from "./mapAndCleanUpdatedInputData.js";
-import lodashMerge from "lodash/merge.js";
-import { removeNullValues, removeUndefinedValues } from "@webiny/utils";
-import { getSystem } from "./system.js";
-import type { SecurityIdentity } from "@webiny/api-core/types/security.js";
-import type { Tenant } from "@webiny/api-core/types/tenancy.js";
-
-interface UpdateEntryFromDataParams<TValues extends CmsEntryValues = CmsEntryValues> {
-    metaInput?: Record<string, any>;
-    model: CmsModel;
-    rawInput: UpdateCmsEntryInput<TValues>;
-    options?: UpdateCmsEntryOptionsInput;
-    context: CmsContext;
-    getIdentity: () => SecurityIdentity;
-    getTenant: () => Tenant;
-    originalEntry: CmsEntry<TValues>;
-}
-
-interface UpdateEntryDataResponse<TValues extends CmsEntryValues = CmsEntryValues> {
-    entry: CmsEntry<TValues>;
-    input: UpdateCmsEntryInput<TValues>;
-}
-
-export const createUpdateEntryData = async <TValues extends CmsEntryValues = CmsEntryValues>({
-    model,
-    rawInput,
-    options,
-    context,
-    metaInput,
-    getIdentity: getSecurityIdentity,
-    originalEntry
-}: UpdateEntryFromDataParams<TValues>): Promise<UpdateEntryDataResponse<TValues>> => {
-    /**
-     * Make sure we only work with fields that are defined in the model.
-     */
-    const cleanedValues = mapAndCleanUpdatedInputData<TValues>(
-        model,
-        rawInput?.values || ({} as TValues)
-    );
-
-    await validateModelEntryDataOrThrow({
-        context,
-        model,
-        values: cleanedValues,
-        entry: originalEntry,
-        skipValidators: options?.skipValidators
-    });
-
-    const mergedValues: TValues = {
-        /**
-         * Existing values from the database, transformed back to original, of course.
-         */
-        ...originalEntry.values,
-        /**
-         * Add new values.
-         */
-        ...cleanedValues
-    };
-
-    const values = await referenceFieldsMapping<TValues>({
-        context,
-        model,
-        values: mergedValues,
-        validateEntries: false
-    });
-
-    /**
-     * If users wants to remove a key from meta values, they need to send meta key with the null value.
-     */
-    const meta = createEntryMeta(metaInput, originalEntry.meta);
-
-    const currentIdentity = getSecurityIdentity();
-    const currentDateTime = new Date();
-
-    /**
-     * We always send the full entry to the hooks and storage operations update.
-     */
-    const entry: CmsEntry<TValues> = {
-        ...originalEntry,
-
-        /**
-         * Revision-level meta fields. 👇
-         */
-        revisionCreatedOn: getDate(rawInput.revisionCreatedOn, originalEntry.revisionCreatedOn),
-        revisionModifiedOn: getDate(rawInput.revisionModifiedOn, currentDateTime),
-        revisionSavedOn: getDate(rawInput.revisionSavedOn, currentDateTime),
-        revisionDeletedOn: getDate(rawInput.revisionDeletedOn, null),
-        revisionRestoredOn: getDate(rawInput.revisionRestoredOn, null),
-        revisionFirstPublishedOn: getDate(
-            rawInput.revisionFirstPublishedOn,
-            originalEntry.revisionFirstPublishedOn
-        ),
-        revisionLastPublishedOn: getDate(
-            rawInput.revisionLastPublishedOn,
-            originalEntry.revisionLastPublishedOn
-        ),
-        revisionCreatedBy: getIdentity(
-            rawInput.revisionCreatedBy,
-            originalEntry.revisionCreatedBy
-        )!,
-        revisionModifiedBy: getIdentity(rawInput.revisionModifiedBy, currentIdentity),
-        revisionSavedBy: getIdentity(rawInput.revisionSavedBy, currentIdentity)!,
-        revisionDeletedBy: getIdentity(rawInput.revisionSavedBy, null),
-        revisionRestoredBy: getIdentity(rawInput.revisionRestoredBy, null),
-        revisionFirstPublishedBy: getIdentity(
-            rawInput.revisionFirstPublishedBy,
-            originalEntry.revisionFirstPublishedBy
-        ),
-        revisionLastPublishedBy: getIdentity(
-            rawInput.revisionLastPublishedBy,
-            originalEntry.revisionLastPublishedBy
-        ),
-
-        /**
-         * Entry-level meta fields. 👇
-         * If required, within storage operations, these entry-level updates
-         * will be propagated to the latest revision too.
-         */
-        createdOn: getDate(rawInput.createdOn, originalEntry.createdOn),
-        savedOn: getDate(rawInput.savedOn, currentDateTime),
-        modifiedOn: getDate(rawInput.modifiedOn, currentDateTime),
-        deletedOn: getDate(rawInput.deletedOn, null),
-        restoredOn: getDate(rawInput.restoredOn, null),
-        firstPublishedOn: getDate(rawInput.firstPublishedOn, originalEntry.firstPublishedOn),
-        lastPublishedOn: getDate(rawInput.lastPublishedOn, originalEntry.lastPublishedOn),
-        createdBy: getIdentity(rawInput.createdBy, originalEntry.createdBy)!,
-        savedBy: getIdentity(rawInput.savedBy, currentIdentity)!,
-        modifiedBy: getIdentity(rawInput.modifiedBy, currentIdentity),
-        deletedBy: getIdentity(rawInput.deletedBy, null),
-        restoredBy: getIdentity(rawInput.restoredBy, null),
-        firstPublishedBy: getIdentity(rawInput.firstPublishedBy, originalEntry.firstPublishedBy),
-        lastPublishedBy: getIdentity(rawInput.lastPublishedBy, originalEntry.lastPublishedBy),
-
-        values,
-        meta,
-        status: transformEntryStatus(originalEntry.status),
-        system: getSystem({
-            input: rawInput,
-            original: originalEntry
-        }),
-        live: originalEntry.live
-    };
-
-    const folderId = rawInput.wbyAco_location?.folderId;
-    if (folderId) {
-        entry.location = {
-            folderId
-        };
-    }
-
-    return {
-        entry,
-        input: {
-            ...rawInput,
-            values: structuredClone(values)
-        }
-    };
-};
-
-/**
- * This method takes original entry meta and new input.
- * When new meta is merged onto the existing one, everything that has undefined or null value is removed.
- */
-const createEntryMeta = (input?: Record<string, any>, original?: Record<string, any>) => {
-    const meta = lodashMerge(original || {}, input || {});
-    return removeUndefinedValues(removeNullValues(meta));
-};
-
-const allowedEntryStatus: string[] = ["draft", "published", "unpublished"];
-
-const transformEntryStatus = (status: CmsEntryStatus | string): CmsEntryStatus => {
-    return allowedEntryStatus.includes(status) ? (status as CmsEntryStatus) : "draft";
-};
