--- packages/api-headless-cms/src/crud/contentEntry/entryDataFactories/index.ts
@@ -1,8 +1,0 @@
-export * from "./createEntryData.js";
-export * from "./createEntryRevisionFromData.js";
-export * from "./createPublishEntryData.js";
-export * from "./createRepublishEntryData.js";
-export * from "./createUnpublishEntryData.js";
-export * from "./createUpdateEntryData.js";
-export * from "./statuses.js";
-export * from "./mapAndCleanUpdatedInputData.js";
