+++ packages/api-headless-cms/src/features/contentEntry/CreateEntry/CreateEntryUseCase.ts
@@ -2,7 +2,7 @@ import { createImplementation, Result } from "@webiny/feature/api";
 import { CreateEntryRepository, CreateEntryUseCase as UseCaseAbstraction } from "./abstractions.js";
 import { EventPublisher } from "@webiny/api-core/features/eventPublisher/index.js";
 import { EntryAfterCreateEvent, EntryBeforeCreateEvent } from "./events.js";
-import { AccessControl, CmsContext } from "~/features/shared/abstractions.js";
+import { AccessControl } from "~/features/shared/abstractions.js";
 import type {
     CmsEntry,
     CmsEntryValues,
@@ -11,27 +11,14 @@ import type {
     CreateCmsEntryOptionsInput
 } from "~/types/index.js";
 import { EntryNotAuthorizedError, EntryValidationError } from "~/domain/contentEntry/errors.js";
-import { createEntryData } from "~/crud/contentEntry/entryDataFactories/createEntryData.js";
-import { TenantContext } from "@webiny/api-core/features/tenancy/TenantContext/index.js";
-import { IdentityContext } from "@webiny/api-core/features/security/IdentityContext/index.js";
+import { CreateEntryDataFactory } from "~/features/contentEntry/entryDataFactories/CreateEntryDataFactory/index.js";
 
-/**
- * CreateEntryUseCase - Orchestrates entry creation.
- *
- * Responsibilities:
- * - Transform raw input to domain entry
- * - Apply access control
- * - Publish domain events
- * - Delegate persistence to repository
- */
 class CreateEntryUseCaseImpl implements UseCaseAbstraction.Interface {
     public constructor(
         private eventPublisher: EventPublisher.Interface,
         private repository: CreateEntryRepository.Interface,
         private accessControl: AccessControl.Interface,
-        private tenantContext: TenantContext.Interface,
-        private identityContext: IdentityContext.Interface,
-        private cmsContext: CmsContext.Interface
+        private createEntryDataFactory: CreateEntryDataFactory.Interface
     ) {}
 
     public async execute<T extends CmsEntryValues = CmsEntryValues>(
@@ -39,25 +26,18 @@ class CreateEntryUseCaseImpl implements UseCaseAbstraction.Interface {
         rawInput: CreateCmsEntryInput<T>,
         options?: CreateCmsEntryOptionsInput
     ): Promise<Result<CmsEntry<T>, UseCaseAbstraction.Error>> {
-        // Check initial access control
         const canAccess = await this.accessControl.canAccessEntry({ model, rwd: "w" });
         if (!canAccess) {
             return Result.fail(EntryNotAuthorizedError.fromModel(model));
         }
 
         try {
-            // Transform raw input to domain entry
-            const { entry, input } = await createEntryData<T>({
+            const { entry, input } = await this.createEntryDataFactory.create<T>(
                 model,
                 rawInput,
-                options,
-                context: this.cmsContext,
-                getIdentity: () => this.identityContext.getIdentity(),
-                getTenant: () => this.tenantContext.getTenant(),
-                accessControl: this.accessControl
-            });
+                options
+            );
 
-            // Apply access control on the created entry
             const canAccessEntry = await this.accessControl.canAccessEntry({
                 model,
                 entry,
@@ -68,16 +48,13 @@ class CreateEntryUseCaseImpl implements UseCaseAbstraction.Interface {
                 return Result.fail(EntryNotAuthorizedError.fromEntry(entry));
             }
 
-            // Publish before event
             await this.eventPublisher.publish(new EntryBeforeCreateEvent({ entry, input, model }));
 
-            // Persist entry
             const result = await this.repository.execute(model, entry);
             if (result.isFail()) {
                 return Result.fail(result.error);
             }
 
-            // Publish after event
             await this.eventPublisher.publish(
                 new EntryAfterCreateEvent({
                     entry,
@@ -91,7 +68,6 @@ class CreateEntryUseCaseImpl implements UseCaseAbstraction.Interface {
             if (error.code === "VALIDATION_FAILED") {
                 return Result.fail(new EntryValidationError(error.message, error.data));
             }
-            // Handle errors from createEntryData or other operations
             return Result.fail(error as UseCaseAbstraction.Error);
         }
     }
@@ -100,12 +76,5 @@ class CreateEntryUseCaseImpl implements UseCaseAbstraction.Interface {
 export const CreateEntryUseCase = createImplementation({
     abstraction: UseCaseAbstraction,
     implementation: CreateEntryUseCaseImpl,
-    dependencies: [
-        EventPublisher,
-        CreateEntryRepository,
-        AccessControl,
-        TenantContext,
-        IdentityContext,
-        CmsContext
-    ]
+    dependencies: [EventPublisher, CreateEntryRepository, AccessControl, CreateEntryDataFactory]
 });
