+++ packages/api-headless-cms/src/features/contentEntry/UpdateEntry/UpdateEntryUseCase.ts
@@ -5,9 +5,6 @@ import { UpdateEntryRepository } from "./abstractions.js";
 import { EventPublisher } from "@webiny/api-core/features/eventPublisher/index.js";
 import { EntryBeforeUpdateEvent, EntryAfterUpdateEvent } from "./events.js";
 import { AccessControl } from "~/features/shared/abstractions.js";
-import { CmsContext } from "~/features/shared/abstractions.js";
-import { TenantContext } from "@webiny/api-core/features/tenancy/TenantContext/index.js";
-import { IdentityContext } from "@webiny/api-core/features/security/IdentityContext/index.js";
 import { GetRevisionByIdUseCase } from "~/features/contentEntry/GetRevisionById/abstractions.js";
 import type {
     CmsEntry,
@@ -19,28 +16,15 @@ import type {
 import type { GenericRecord } from "@webiny/api/types.js";
 import { EntryNotAuthorizedError } from "~/domain/contentEntry/errors.js";
 import { EntryLockedError } from "~/domain/contentEntry/errors.js";
-import { createUpdateEntryData } from "~/crud/contentEntry/entryDataFactories/createUpdateEntryData.js";
+import { UpdateEntryDataFactory } from "~/features/contentEntry/entryDataFactories/UpdateEntryDataFactory/index.js";
 
-/**
- * UpdateEntryUseCase - Orchestrates entry updates.
- *
- * Responsibilities:
- * - Fetch original entry
- * - Validate entry is not locked
- * - Transform raw input to updated domain entry
- * - Apply access control
- * - Publish domain events
- * - Delegate persistence to repository
- */
 class UpdateEntryUseCaseImpl implements UseCaseAbstraction.Interface {
     public constructor(
         private eventPublisher: EventPublisher.Interface,
         private repository: UpdateEntryRepository.Interface,
         private accessControl: AccessControl.Interface,
-        private cmsContext: CmsContext.Interface,
-        private tenantContext: TenantContext.Interface,
-        private identityContext: IdentityContext.Interface,
-        private getRevisionByIdUseCase: GetRevisionByIdUseCase.Interface
+        private getRevisionByIdUseCase: GetRevisionByIdUseCase.Interface,
+        private updateEntryDataFactory: UpdateEntryDataFactory.Interface
     ) {}
 
     async execute<T extends CmsEntryValues = CmsEntryValues>(
@@ -50,7 +34,6 @@ class UpdateEntryUseCaseImpl implements UseCaseAbstraction.Interface {
         metaInput?: GenericRecord,
         options?: UpdateCmsEntryOptionsInput
     ): Promise<Result<CmsEntry<T>, UseCaseAbstraction.Error>> {
-        // Check initial access control
         const canAccess = await this.accessControl.canAccessEntry({ model, rwd: "w" });
         if (!canAccess) {
             return Result.fail(EntryNotAuthorizedError.fromModel(model));
@@ -65,24 +48,18 @@ class UpdateEntryUseCaseImpl implements UseCaseAbstraction.Interface {
 
             const originalEntry = result.value;
 
-            // Check if entry is locked
             if (originalEntry.locked) {
                 return Result.fail(new EntryLockedError());
             }
 
-            // Transform raw input to updated domain entry
-            const { entry, input } = await createUpdateEntryData<T>({
+            const { entry, input } = await this.updateEntryDataFactory.create<T>(
                 model,
                 rawInput,
-                options,
-                context: this.cmsContext,
-                getIdentity: () => this.identityContext.getIdentity(),
-                getTenant: () => this.tenantContext.getTenant(),
                 originalEntry,
+                options,
                 metaInput
-            });
+            );
 
-            // Apply access control on the updated entry
             const canAccessEntry = await this.accessControl.canAccessEntry({
                 model,
                 entry,
@@ -93,18 +70,15 @@ class UpdateEntryUseCaseImpl implements UseCaseAbstraction.Interface {
                 return Result.fail(EntryNotAuthorizedError.fromModel(model));
             }
 
-            // Publish before event
             await this.eventPublisher.publish(
                 new EntryBeforeUpdateEvent({ entry, original: originalEntry, input, model })
             );
 
-            // Persist updated entry
             const updateResult = await this.repository.execute(model, entry);
             if (updateResult.isFail()) {
                 return Result.fail(updateResult.error);
             }
 
-            // Publish after event
             await this.eventPublisher.publish(
                 new EntryAfterUpdateEvent({
                     entry,
@@ -116,7 +90,6 @@ class UpdateEntryUseCaseImpl implements UseCaseAbstraction.Interface {
 
             return Result.ok(entry);
         } catch (error) {
-            // Handle errors from createUpdateEntryData or other operations
             return Result.fail(error as UseCaseAbstraction.Error);
         }
     }
@@ -129,9 +102,7 @@ export const UpdateEntryUseCase = createImplementation({
         EventPublisher,
         UpdateEntryRepository,
         AccessControl,
-        CmsContext,
-        TenantContext,
-        IdentityContext,
-        GetRevisionByIdUseCase
+        GetRevisionByIdUseCase,
+        UpdateEntryDataFactory
     ]
 });
