+++ packages/api-headless-cms/src/features/contentEntry/entryDataFactories/CreateRepublishEntryDataFactory/CreateRepublishEntryDataFactory.ts
@@ -6,8 +6,11 @@ import {
 } from "./abstractions.js";
 import { CmsContext } from "~/features/shared/abstractions.js";
 import { IdentityContext } from "@webiny/api-core/features/security/IdentityContext/index.js";
-import { createRepublishEntryData } from "~/crud/contentEntry/entryDataFactories/createRepublishEntryData.js";
 import type { CmsEntry, CmsEntryValues, CmsModel } from "~/types/index.js";
+import { referenceFieldsMapping } from "~/crud/contentEntry/referenceFieldsMapping.js";
+import { STATUS_PUBLISHED } from "../statuses.js";
+import { getIdentity } from "~/utils/identity.js";
+import { getDate } from "~/utils/date.js";
 
 class CreateRepublishEntryDataFactoryImpl implements ICreateRepublishEntryDataFactory {
     public constructor(
@@ -15,16 +18,52 @@ class CreateRepublishEntryDataFactoryImpl implements ICreateRepublishEntryDataFa
         private readonly identityContext: IdentityContext.Interface
     ) {}
 
-    public create<TValues extends CmsEntryValues = CmsEntryValues>(
+    public async create<TValues extends CmsEntryValues = CmsEntryValues>(
         model: CmsModel,
         originalEntry: CmsEntry<TValues>
     ): Promise<ICreateRepublishEntryDataResponse<TValues>> {
-        return createRepublishEntryData<TValues>({
-            model,
-            originalEntry,
+        const values = await referenceFieldsMapping<TValues>({
             context: this.cmsContext,
-            getIdentity: () => this.identityContext.getIdentity()
+            model,
+            values: originalEntry.values,
+            validateEntries: false
         });
+
+        const currentDateTime = new Date().toISOString();
+        const currentIdentity = this.identityContext.getIdentity();
+
+        const entry: CmsEntry<TValues> = {
+            ...originalEntry,
+            status: STATUS_PUBLISHED,
+            savedOn: getDate(currentDateTime),
+            modifiedOn: getDate(currentDateTime),
+            savedBy: getIdentity(currentIdentity)!,
+            modifiedBy: getIdentity(currentIdentity),
+            firstPublishedOn: getDate(originalEntry.firstPublishedOn, currentDateTime),
+            firstPublishedBy: getIdentity(originalEntry.firstPublishedBy, currentIdentity),
+            lastPublishedOn: getDate(currentDateTime),
+            lastPublishedBy: getIdentity(currentIdentity),
+            revisionSavedOn: getDate(currentDateTime),
+            revisionModifiedOn: getDate(currentDateTime),
+            revisionSavedBy: getIdentity(currentIdentity)!,
+            revisionModifiedBy: getIdentity(currentIdentity),
+            revisionFirstPublishedOn: getDate(
+                originalEntry.revisionFirstPublishedOn,
+                currentDateTime
+            ),
+            revisionFirstPublishedBy: getIdentity(
+                originalEntry.revisionFirstPublishedBy,
+                currentIdentity
+            ),
+            revisionLastPublishedOn: getDate(currentDateTime),
+            revisionLastPublishedBy: getIdentity(currentIdentity),
+            values,
+            live: {
+                version: originalEntry.version
+            }
+        };
+
+        return { entry };
     }
 }
