+++ packages/api-headless-cms/src/features/contentEntry/entryDataFactories/CreateUnpublishEntryDataFactory/CreateUnpublishEntryDataFactory.ts
@@ -5,19 +5,35 @@ import {
     type ICreateUnpublishEntryDataResponse
 } from "./abstractions.js";
 import { IdentityContext } from "@webiny/api-core/features/security/IdentityContext/index.js";
-import { createUnpublishEntryData } from "~/crud/contentEntry/entryDataFactories/createUnpublishEntryData.js";
 import type { CmsEntry, CmsEntryValues } from "~/types/index.js";
+import { STATUS_UNPUBLISHED } from "../statuses.js";
+import { getIdentity } from "~/utils/identity.js";
+import { getDate } from "~/utils/date.js";
 
 class CreateUnpublishEntryDataFactoryImpl implements ICreateUnpublishEntryDataFactory {
     public constructor(private readonly identityContext: IdentityContext.Interface) {}
 
-    public create<TValues extends CmsEntryValues = CmsEntryValues>(
+    public async create<TValues extends CmsEntryValues = CmsEntryValues>(
         originalEntry: CmsEntry<TValues>
     ): Promise<ICreateUnpublishEntryDataResponse<TValues>> {
-        return createUnpublishEntryData<TValues>({
-            originalEntry,
-            getIdentity: () => this.identityContext.getIdentity()
-        });
+        const currentDateTime = new Date().toISOString();
+        const currentIdentity = this.identityContext.getIdentity();
+
+        const entry: CmsEntry<TValues> = {
+            ...originalEntry,
+            status: STATUS_UNPUBLISHED,
+            savedOn: getDate(currentDateTime),
+            modifiedOn: getDate(currentDateTime),
+            savedBy: getIdentity(currentIdentity),
+            modifiedBy: getIdentity(currentIdentity),
+            revisionSavedOn: getDate(currentDateTime),
+            revisionModifiedOn: getDate(currentDateTime),
+            revisionSavedBy: getIdentity(currentIdentity),
+            revisionModifiedBy: getIdentity(currentIdentity),
+            live: null
+        };
+
+        return { entry };
     }
 }
