+++ packages/api-headless-cms/src/features/contentEntry/entryDataFactories/mapAndCleanUpdatedInputData.ts
@@ -1,9 +1,6 @@
 import WebinyError from "@webiny/error";
 import type { CmsEntryValues, CmsModel } from "~/types/index.js";
 
-/**
- * Cleans the update input entry data.
- */
 export const mapAndCleanUpdatedInputData = <TValues extends CmsEntryValues = CmsEntryValues>(
     model: CmsModel,
     input?: Partial<TValues>
@@ -12,18 +9,14 @@ export const mapAndCleanUpdatedInputData = <TValues extends CmsEntryValues = Cms
         return {};
     }
     return model.fields.reduce<Partial<TValues>>((acc, field) => {
-        /**
-         * This should never happen, but let's make it sure.
-         * The fix would be for the user to add the fieldId on the field definition.
-         */
+        /* This should never happen, but let's make it sure.
+         * The fix would be for the user to add the fieldId on the field definition. */
         if (!field.fieldId) {
             throw new WebinyError("Field does not have an fieldId.", "MISSING_FIELD_ID", {
                 field
             });
         }
-        /**
-         * We cannot set default value here because user might want to update only certain field values.
-         */
+        /* We cannot set default value here because user might want to update only certain field values. */
         const key = field.fieldId as keyof TValues;
         const value = input[key];
         if (value === undefined) {
